requirements.txt
certifi==2020.12.5
docxtpl==0.11.3
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.2
python-docx==0.8.10
six==1.15.0
streamlit==1.10.0
